+++ packages/agtree/src/converter/cosmetic/scriptlet.ts
@@ -292,7 +292,10 @@ export class ScriptletRuleConverter extends RuleConverterBase {
             } else {
                 // Otherwise, try to find the corresponding uBO scriptlet name, or use the original one if not found
                 const uboScriptlet = scriptletsCompatibilityTable.getFirst(scriptletName, GenericPlatform.UboAny);
-                uboScriptletName = uboScriptlet?.name ?? scriptletName;
+                if (!uboScriptlet) {
+                    throw new Error(`Scriptlet "${scriptletName}" is not supported in uBlock Origin.`);
+                }
+                uboScriptletName = uboScriptlet.name;
             }
 
             // Remove the '.js' suffix if it's there - its presence is not mandatory