+++ packages/agtree/CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Changed
+
+- Limit conversion of `ADG` scriptlets to `uBO` if no equivalents are available.
+
+[Unreleased]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/agtree
+
 ## [3.1.0] - 2025-03-31
 
 ### Added